Restructure python modules for user admin + get partial serializer working so core editting features now fully work.

+++ premium/web-frontend/modules/baserow_premium/services/userAdmin.js
@@ -1,13 +1,14 @@
 export default (client) => {
   return {
     fetchPage(page) {
-      return client.get(`/premium/user/?page=${page}`)
+      return client.get(`/premium/admin/user/?page=${page}`)
     },
     update(userId, values) {
-      return client.patch(`/premium/user/${userId}/`, values)
+      values.id = userId
+      return client.patch(`/premium/admin/user/${userId}/`, values)
     },
     delete(userId) {
-      return client.delete(`/premium/user/${userId}/`)
+      return client.delete(`/premium/admin/user/${userId}/`)
     },
   }
 }
